Metadata middleware basePath matching
swagger-metadata is creating the path regexps properly but when it goes
to find one that matches the current path (swagger-megadata.js line 98)
it doesn't take the basePath into account.

This patch parses the path part of the basePath of each resource and
prepends it to the generated api regexps.

Fixes #13