Get mock values from `schema.example` (if no `default` value)
In many cases you don't want to define a `default` value for a schema
object, but you still want to provide representative examples,
particularly in mock responses.

This is particularly important where `string` type uses a `pattern` that
doesn't match the hard-coded mock values, which in turns causes the
mocked response to fail the response validator. This change allows mock
values to be pulled from the `example` field to solve this problem.

`example` is inserted between `default` and `enum`, leaving the the
mock value source precedence as:
1. `1.2` `defaultValue`
2. `2.0` `default`
3. `2.0` `example`
4. `enum` (first value)
5. hardcoded value (or *now* for date/date-time)

Test plan:
* Added unit tests to cover the appropriate selection of `default` and `example` as the mock value for string, number, integer, and boolean types.
* Used gulp to run lint and all unit tests and ensured everything passes.
* Used coverage report to verify all new lines were tested at least once.