Do not coerce non-array bodies into an array
If you ever had a `body` parameter whose type was `array` but did not
provide an `array` to the request, we were previously coercing the body
into a single-item array.  The reason for this was to handle query
parameters that were type array but only a single value was provided.

This commit makes it so that body parameters never behave this way.

Fixes #438