Adds options.invalidResponseHook to swagger-validator
If provided, the invalidResponseHook middleware will be called when
response validation fails. A swagger-validator error will be passed
in err; the middleware may choose to propogate or ignore the error.

Call next(err) to let express handle the error normally. Call next()
to quash the error and continue through the normal middleware
stack.