Upgrade from lodash-compat to lodash v4
The `isFunction()` function in lodash v3 does not properly recognize
`async` functions in Node.js v7 or later. This is fixed in lodash v4.
Fortunately, the only breaking change that seemed to affect
swagger-tools is the change to module paths for requiring individual
functions.

The lodash-compat module that was used has been deprecated and is no
longer supported. Fortunately, plain lodash seems to work just fine.

Fixes #443